addressed changes
